refactor : Image API Logic 구성 View -> ViewModel -> Model -> Repository 로 요청하여 결과 반영까지 완료
