+++ ImageSearch/Sources/Entity/Items.swift
@@ -1,0 +1,13 @@
+//
+//  Image.swift
+//  ImageSearch
+//
+//  Created by once on 2021/06/21.
+//
+
+import Foundation
+
+struct Items: Codable {
+  let documents: [Image]
+  let meta: Meta
+}
