+++ ImageSearch/Sources/Repository/NetworkService.swift
@@ -31,7 +31,23 @@ class NetworkService: NetworkServiceType {
 
 extension NetworkService {
   private func getImageList(paramData: String, onComplete: @escaping (Result<Data, Error>) -> Void) {
-    let param = ["query": paramData]
+    let param = ["query": "사탕"]
+    
+    
+//    NetworkManager
+//      .shared
+//      .session
+//      .request(NetworkRouter.searchImage(params: param))
+//      .responseJSON { response in
+//        switch response.result {
+//        case let .success(data):
+//          print(data)
+//          onComplete(.success(data))
+//        case let .failure(error):
+//          print(error)
+//          onComplete(.failure(error))
+//        }
+//      }
     
     NetworkManager
       .shared
